Two different configuration files: config.toml and --options FILE
The first contains generic configuration arguments and is a TOML file,
the second is optional and is called from command line. It contains
command line arguments in the style of boost program option file.

The main difference is that the boost one does not need double quotes,
cannot manage spaces and can contain multiple definition of the same
scalar.