Multiple algos for identities
It is now possible to define more algos that will be used for signature and
identity. T the moment the only implemented algorithms are "bitcoin" and
"dummy".

The config file contains the option on what algos are in use

Identity class checks if the identity key is well-formed against algos in use

The Signature class checks the same for the signature of a node