Primary key index on current HashSet
Add is slow {#32}
---
**closed**
***internals,db***

At the moment there is an index that is calculated every time. This should be the cause. {#31} can be tryied to see how slow it is.

Maybe we should add a _import_ command that accepts multiple nodes, using the index only at the end of the insert.

Some benchmark on 172 nodes of issue #31:
* In commit "hash-object command" the time was 23 minutes
* In the next commit the time is 6 seconds

Change status to **closed**